Complete encoding of periodical variants.
All known variants from texts are included in this encoding. The
encoding still needs to be standardized so that certain things are
marked consistently, etc etc. Moreover, important texts (introductions,
poems appearing in periodicals concurrently with poems that were later
integrated into *Harlem Shadows*) are currently **not** included.
Including this latter will likely involve a reorganization of the file
and its encoding, no longer as a facsimile edition of the 1922 *Harlem
Shadows* plus textual variants, but as a new critical text.